Changed interface for Model.DrawFrame() method.
Now takes 'time' input as a float rather than an int. This allows
for smoothly slowing down animation to an arbitrary degree.

+++ include/Md5Model.h
@@ -45,7 +45,7 @@ namespace asteria {
          virtual bool UnstageModel( void );
 
          // Draw animation 'anim' at 'time' milliseconds.
-         virtual bool DrawFrame( int anim, int time );
+         virtual bool DrawFrame( int anim, float time );
 
          // Draws the base mesh.
          virtual bool DrawStatic( void );
